🔧 Replace csscomb with prettier
- Remove `csscomb` and its config;
- Add Prettier as a StyleLint plugin along with its configuration;

CSSComb seems as it is not currently maintained, with lots of issues on its GitHub repo, as well as not properly functioning. the simplest way out of this is to rely on already present StyleLint package which can automatically fix certain aspects of CSS formatting via `--fix` option.

Using Prettier no-brainer config we don’t have to worry about any stylistic problems and can just focus on writing maintainable code.